Added two new courses to the list
Hi there! Two new courses have been added to the list:

* Git and Github for beginners (Loiane)
* Angular + ASP.NET Core API + EF Core

Note: Regarding the git course, it was produced by Loiane Groner and has the same title as the first course on the list (produced by Willian Justen), but they are different contents. I thought it was better to differentiate it by putting the author's name in front.